Use `module-alias` to improve module resolution in Brigade scripts
Previously, local dependency resolution and Brigadier aliasing were done
by a special `require` replacement, which was only available in the main
Brigade script (eg, `brigade.js`).  This meant that local dependencies
could not themselves depend on packages installed via `brigade.json`.

Now, `module-alias` takes care of Brigadier aliasing and applies the
correct module resolution paths to allow local dependencies to require
packages from the `node_modules` that `brigade.json` targets.

Also:
- Updated dependencies.md with new module resolution semantics
- Removed obsolete `require` module and tests
- Removed obsolete `brigade.js` file
