--- brigade-worker/src/brigade.js
@@ -1,2 +1,0 @@
-// This file is here to satisfy the tsc build.
-// Upon startup, the real brigade.js replaces it.
